# api/openapi.yaml
openapi: 3.1.0
info:
  title: Centipede API
  version: 0.2.0
  description: 阅读型多语言词汇本 API。业务数据只通过 Centipede 本地用户隔离。
servers:
  - url: http://localhost:7788
paths:
  /health/live:
    get:
      operationId: healthLive
      responses:
        '200': {description: Process is alive}
  /health/ready:
    get:
      operationId: healthReady
      responses:
        '200': {description: Process and PostgreSQL are ready}
        '503': {description: A required dependency is unavailable}
  /api/v1/system/info:
    get:
      operationId: systemInfo
      responses:
        '200':
          description: Application metadata
          content: {application/json: {schema: {$ref: '#/components/schemas/SystemInfo'}}}
  /api/v1/auth/register:
    post:
      operationId: register
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Credentials'}}}}
      responses:
        '200': {description: Account and access token returned; refresh token is an HttpOnly cookie}
        '400': {description: Invalid registration data}
        '409': {description: Email already exists}
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/Credentials'}}}}
      responses:
        '200': {description: Account and access token returned; refresh token is an HttpOnly cookie}
        '401': {description: Invalid credentials}
  /api/v1/auth/refresh:
    post:
      operationId: refresh
      responses:
        '200': {description: Rotated access and refresh tokens}
        '401': {description: Refresh cookie is invalid or expired}
  /api/v1/auth/logout:
    post:
      operationId: logout
      responses: {'204': {description: Session revoked}}
  /api/v1/auth/me:
    get:
      operationId: currentUser
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Current local user, content: {application/json: {schema: {$ref: '#/components/schemas/UserResponse'}}}}
        '401': {description: Invalid or missing access token}
  /api/v1/vocabulary/entries:
    post:
      operationId: createVocabularyEntry
      security: [{bearerAuth: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyEntryInput'}}}}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/VocabularyEntry'}}}}
        '400': {description: Invalid entry}
        '401': {description: Authentication required}
    get:
      operationId: listVocabularyEntries
      security: [{bearerAuth: []}]
      parameters:
        - {in: query, name: language, schema: {type: string}, description: BCP-47 style language tag}
        - {in: query, name: status, schema: {type: string, enum: [new, learning, known, paused]}}
        - {in: query, name: q, schema: {type: string}, description: Search original text or definition}
        - {in: query, name: limit, schema: {type: integer, default: 50, maximum: 100}}
        - {in: query, name: offset, schema: {type: integer, default: 0, minimum: 0}}
      responses:
        '200': {description: Entries owned by the current user}
        '401': {description: Authentication required}
  /api/v1/vocabulary/entries/{id}:
    get:
      operationId: getVocabularyEntry
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: id, required: true, schema: {type: integer, format: int64}}]
      responses:
        '200': {description: Vocabulary entry}
        '404': {description: Entry not found}
  /api/v1/vocabulary/entries/{id}/status:
    patch:
      operationId: updateVocabularyStatus
      security: [{bearerAuth: []}]
      parameters: [{in: path, name: id, required: true, schema: {type: integer, format: int64}}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties: {status: {type: string, enum: [new, learning, known, paused]}}
      responses:
        '200': {description: Updated vocabulary entry}
        '400': {description: Invalid status}
        '404': {description: Entry not found}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    SystemInfo:
      type: object
      properties: {name: {type: string}, architecture: {type: string}, identity_provider: {type: string}}
    Credentials:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string, minLength: 8, maxLength: 128}
        display_name: {type: string, maxLength: 80}
    User:
      type: object
      properties:
        id: {type: integer, format: int64}
        email: {type: string}
        display_name: {type: string}
        status: {type: string}
    UserResponse:
      type: object
      properties: {user: {$ref: '#/components/schemas/User'}}
    Source:
      type: object
      properties:
        book: {type: string}
        chapter: {type: string}
        location: {type: string}
    Context:
      type: object
      required: [text]
      properties:
        id: {type: integer, format: int64}
        text: {type: string}
        location: {type: string}
    VocabularyEntryInput:
      type: object
      required: [language, text]
      properties:
        language: {type: string, description: BCP-47 style language tag, example: en-US}
        text: {type: string, maxLength: 1000}
        lemma: {type: string}
        definition: {type: string}
        notes: {type: string}
        status: {type: string, enum: [new, learning, known, paused], default: new}
        source: {$ref: '#/components/schemas/Source'}
        contexts: {type: array, items: {$ref: '#/components/schemas/Context'}}
    VocabularyEntry:
      allOf:
        - {$ref: '#/components/schemas/VocabularyEntryInput'}
        - type: object
          properties:
            id: {type: integer, format: int64}
            normalized_text: {type: string}
            created_at: {type: string, format: date-time}
            updated_at: {type: string, format: date-time}